% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRFunctions.R
\docType{data}
\name{average_intensity}
\alias{average_intensity}
\title{This function calculates the average intensity and prints it.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
average_intensity
}
\description{
This function calculates the average intensity and prints it.
}
\keyword{datasets}
