% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRFunctions.R
\docType{data}
\name{average_wavenumber}
\alias{average_wavenumber}
\title{This function calculates the average wavenumber and prints it.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
average_wavenumber
}
\description{
This function calculates the average wavenumber and prints it.
}
\keyword{datasets}
