% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_wavenumber.R
\name{calculate_average_wavenumber}
\alias{calculate_average_wavenumber}
\title{Calculate Average Wavenumber}
\usage{
calculate_average_wavenumber(raman_hdpe)
}
\arguments{
\item{raman_hdpe}{Raman spectra data; feel free to upload your own data instead}
}
\value{
Average wavenumber of the entire data set
}
\description{
This function calculates the average wavenumber of a given Raman spectra data set.
}
\examples{
data(raman_hdpe, package = "OpenSpecy")
calculate_average_wavenumber(raman_hdpe)

}
