% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRFunctions.R
\name{print_average_wavenumber}
\alias{print_average_wavenumber}
\title{This function prints the average wavenumber.}
\usage{
print_average_wavenumber(average_wavenumber)
}
\description{
This function prints the average wavenumber.
}
