% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRFunctions.R
\name{calculate_average_intensity}
\alias{calculate_average_intensity}
\title{This function calculates the average intensity.}
\usage{
calculate_average_intensity(raman_hdpe)
}
\description{
This function calculates the average intensity.
}
