% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_intensity.R
\name{calculate_average_intensity}
\alias{calculate_average_intensity}
\title{Calculate Average Intensity}
\usage{
calculate_average_intensity(raman_hdpe)
}
\arguments{
\item{raman_hdpe}{Raman spectra data set; can substitute with your own data set}
}
\value{
Average intensity of the entire data set
}
\description{
Calculate Average Intensity
}
\examples{
calculate_average_intensity <- function(your_data) `...`

}
