% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_wavenumber.R
\name{calculate_average_wavenumber}
\alias{calculate_average_wavenumber}
\title{Calculate Average Wavenumber}
\usage{
calculate_average_wavenumber(raman_hdpe)
}
\arguments{
\item{raman_hdpe}{Raman spectra data; can substitute with different data set}
}
\value{
Average wavenumber of an entire data set
}
\description{
Calculate Average Wavenumber
}
\examples{
calculate_average_wavenumber <- function(your_data) `...`

}
