% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_wavenumber.R
\name{calculate_average_wavenumber}
\alias{calculate_average_wavenumber}
\title{Calculating average wavenumber}
\usage{
calculate_average_wavenumber(raman_hdpe)
}
\arguments{
\item{raman_hdpe}{}
}
\value{
Printing the average wavenumber
}
\description{
Calculating average wavenumber
}
\examples{
Values from the raman_hdpe data set
}
