% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRFunctions.R
\name{calculate_average_wavenumber}
\alias{calculate_average_wavenumber}
\title{This function calculates the average wavenumber. If there's no data inputted, an automatic response will be generated.}
\usage{
calculate_average_wavenumber(raman_hdpe)
}
\description{
This function calculates the average wavenumber. If there's no data inputted, an automatic response will be generated.
}
