% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_intensity_ratio.R
\name{calculate_intensity_ratio}
\alias{calculate_intensity_ratio}
\title{Calculate Intensity Ratio}
\usage{
calculate_intensity_ratio(peak1_intensity, peak2_intensity)
}
\arguments{
\item{peak1_intensity}{The intensity of the first peak}

\item{peak2_intensity}{The intensity of the second peak}
}
\value{
Ratio between the two intensities
}
\description{
Calculate Intensity Ratio
}
\examples{
calculate_intensity_ratio(20, 30)

}
