% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRFunctions.R
\name{calculate_intensity_ratio}
\alias{calculate_intensity_ratio}
\title{My Function 3
This function calculates the intensity ratio between two peaks of the user's choice.}
\usage{
calculate_intensity_ratio(peak1_intensity, peak2_intensity)
}
\description{
My Function 3
This function calculates the intensity ratio between two peaks of the user's choice.
}
