% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_wavenumber.R
\name{print_average_wavenumber}
\alias{print_average_wavenumber}
\title{Print Average Wavenumber}
\usage{
print_average_wavenumber(average_wavenumber)
}
\arguments{
\item{average_wavenumber}{The average wavenumber to be printed}
}
\description{
This function prints the average wavenumber to the console.
}
